create order (no payment yet)

+++ oliolishop/src/main/java/com/oliolishop/oliolishop/dto/customer/CustomerResponse.java
@@ -4,6 +4,7 @@ package com.oliolishop.oliolishop.dto.customer;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -16,6 +17,6 @@ public class CustomerResponse {
     String name;
     LocalDate dob;
     String gender;
-    int loyaltyPoints;
+    BigDecimal loyaltyPoints;
     String image;
 }
